requirements.txt
opencv-python
tqdm
torch
torchvision
numpy
matplotlib
scikit-learn
pillow
psutil